chore: editor 하위 컴포넌트 분리
