+++ src/views/note/note-form/Editor.tsx
@@ -6,8 +6,6 @@ import { Controller, useFormContext } from "react-hook-form";
 
 import { useModalContext } from "@/contexts/InputModalContext";
 
-import LinkDisplay from "./LinkDisplay";
-
 const ReactQuill = dynamic(() => import("react-quill-new"), { ssr: false });
 
 const toolbarOptions = [
@@ -20,9 +18,7 @@ const toolbarOptions = [
 
 export default function Editor() {
   const { openModal } = useModalContext();
-  const { watch, control } = useFormContext();
-
-  const content = watch("content");
+  const { control } = useFormContext();
 
   const modules = useMemo(
     () => ({
@@ -37,25 +33,18 @@ export default function Editor() {
   );
 
   return (
-    <div className="flex flex-1 flex-col gap-2">
-      <p className="pt-[12px] text-xs font-medium text-slate-800">
-        공백포함 : 총 {content?.length || 0}자 | 공백제외 : 총{" "}
-        {content?.replace(/\s/gm, "").length || 0}자
-      </p>
-      <LinkDisplay />
-      <Controller
-        control={control}
-        name="content"
-        render={({ field: { onChange, value } }) => (
-          <ReactQuill
-            theme="snow"
-            modules={modules}
-            onChange={onChange}
-            value={value}
-            className="flex flex-1 flex-col"
-          ></ReactQuill>
-        )}
-      />
-    </div>
+    <Controller
+      control={control}
+      name="content"
+      render={({ field: { onChange, value } }) => (
+        <ReactQuill
+          theme="snow"
+          modules={modules}
+          onChange={onChange}
+          value={value}
+          className="flex flex-1 flex-col"
+        ></ReactQuill>
+      )}
+    />
   );
 }
